Carrying out some checks.

+++ src/services/categorySevices/DeleteCategoryService.ts
@@ -1,12 +1,19 @@
 import { getRepository } from "typeorm";
 import { Category } from "../../entities/Category";
+import { Video } from "../../entities/Video";
 
 export class DeleteCategoryService {
     async execute(categoryId: string){
         const repo = getRepository(Category);
+        const repoVideo = getRepository(Video);
+
         if(!await repo.findOne(categoryId)){
             return new Error("Category does not exists!");
         }
-        await repo.delete(categoryId);
+
+        if(!await repoVideo.findOne(categoryId)){
+            return new Error("The category is being used by videos!");
+        }
+        await repo.delete(categoryId);        
     }
 }
